Update polkit configuration and add systemd service
This commit updates the polkit configuration in the `default.nix` file for the `albus` host. It adds a new polkit rule that allows users in the "power" group to perform certain actions. Additionally, a new systemd service `polkit-gnome-authentication-agent-1` is added to the `systemd` section. This service is configured to start with the graphical session and restart on failure.